+++ CHANGELOG.md
@@ -24,6 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [#916](https://github.com/spegel-org/spegel/pull/916) Refactor OCI client options and add header configuration.
 - [#931](https://github.com/spegel-org/spegel/pull/931) Cleanup e2e tests to use more standardized node labels.
 - [#932](https://github.com/spegel-org/spegel/pull/932) Switch to using spegel-org images in e2e tests.
+- [#957](https://github.com/spegel-org/spegel/pull/957) Refactor memory store to include descriptor.
 
 ### Deprecated